[chore] Fixed performance issues generating the chain lists
This commit makes the assumption that the latest stat shouldn't be too
far outside the stat range. By moving the where clause containing the
stat range inside the main where clause, the query will not use the
latest stat of users but the latest stat within that range. But this
change to the query significantly decreases the number of rows the query
will "operate" upon.

Signed-off-by: tiksan <webmaster@deek.sh>